feat: implemented minimal requests to frontend
