+++ packages/angular/src/app/app.routes.ts
@@ -1,7 +1,4 @@
 import { Routes } from '@angular/router';
 import { MainComponent } from './pages/main/main.component';
-import { ProjectsListComponent } from './components/projects-list/projects-list.component';
 
-export const routes: Routes = [
-  { path: '', component: MainComponent },
-];
+export const routes: Routes = [{ path: '', component: MainComponent }];
