+++ packages/angular/src/app/components/header/header.component.ts
@@ -2,6 +2,7 @@ import { Component, CUSTOM_ELEMENTS_SCHEMA, Input, OnInit } from '@angular/core'
 import { MatIconModule } from '@angular/material/icon';
 import { ApiService } from '../../services/api.service';
 import { ActivatedRoute } from '@angular/router';
+import { Project } from '../../models/kanban.model';
 
 @Component({
   selector: 'app-header',
@@ -12,19 +13,23 @@ import { ActivatedRoute } from '@angular/router';
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
 })
 export class HeaderComponent implements OnInit {
-  title?: string = '';
-  id: number = 0;
-  constructor(
-    private apiService: ApiService,
-    private route: ActivatedRoute,
-  ) {}
+  getTitle: any;
+  title!: string;
+  project_id!: string;
+
+  constructor(private apiService: ApiService, private route: ActivatedRoute) {}
 
   ngOnInit(): void {
-    this.route.queryParams.subscribe((value) => (this.id = value['id']));
-    this.getProjectDataById(this.id);
-  }
+    this.apiService.getAllProjects().subscribe((projects) => {
+      this.getTitle = projects;
+      this.title = this.getTitle[0].title;
+    });
 
-  getProjectDataById(id: number): void {
-    this.apiService.getProjectsDataApiById(id).subscribe((res) => (this.title = res.title));
+    this.route.queryParams.subscribe((value) => {
+      this.project_id = value['project_id'];
+      if (this.project_id) {
+        this.apiService.getProjectById(this.project_id).subscribe((project) => (this.title = project.title));
+      }
+    });
   }
 }
