+++ packages/angular/src/app/models/kanban.model.ts
@@ -1,34 +1,15 @@
-export interface Kanban {
-  titleProject: string;
-  columns: [
-    {
-      titleColumn: string;
-      tasks: [
-        {
-          titleTask: string;
-          description: string;
-        },
-      ];
-    },
-  ];
-}
-
 export interface Project {
-  id: number;
   title: string;
-  columns: Array<Column>;
 }
 
 export interface Column {
-  id: number;
-  id_project: number;
+  _id_project: string;
   title: string;
-  tasks: Array<Task>;
-  excluded: boolean;
 }
 
 export interface Task {
-  id: number;
+  _id_project: string;
+  _id_column: string;
   title: string;
   description: string;
   archived: boolean;
