+++ packages/angular/src/environments/environment.development.ts
@@ -1,4 +1,4 @@
 export const environment = {
   production: true,
-  apiUrl: 'http://localhost:3000/',
+  apiUrl: 'http://localhost:3000',
 };
